MediaFiles interface for Movies and Series
